migrated new feedback & contact endpoints to v4
